Moved the code for PhysicalCTETable to the physicalop package.
